requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.3
pytest==8.3.4
httpx>=0.25.0,<0.28